Add Redis Healthcheck  & Refactor Health check (#216)
* fix sql healthcheck
* add redis health check
* remove health handler test as it is not checking functionality

---------
